backend-python/video-processor/requirements.txt
fastapi
uvicorn
boto3
ffmpeg-python
python-multipart
requests
python-dotenv